添加功能页面和业务类

+++ src/main/java/com/dyenigma/dao/BaseMapper.java
@@ -7,9 +7,9 @@ public interface BaseMapper<T> {
 
     int update(T t);
 
-    int delete(long id);
+    int delete(int id);
 
-    T findById(long id);
+    T findById(int id);
 
     List<T> findAll();
 }
