+++ src/main/java/com/dyenigma/service/BaseService.java
@@ -8,9 +8,9 @@ public interface BaseService<T> {
 
     int update(T t);
 
-    int delete(long id);
+    int delete(int id);
 
-    T findById(long id);
+    T findById(int id);
 
     List<T> findAll();
 }
